requirements.txt
asgiref==3.5.2
attrs==22.1.0
autobahn==22.6.1
Automat==20.2.0
backports.zoneinfo==0.2.1
boto3==1.24.26
botocore==1.27.26
certifi==2021.10.8
cffi==1.15.0
channels==3.0.5
chardet==4.0.0
charset-normalizer==2.0.12
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
daphne==3.0.2
Django==4.1
django-cte==1.2.1
django-debug-toolbar==3.2.4
django-filter==22.1
django-rest-knox==4.1.0
django-reversion==5.0.1
django-storages==1.12.3
djangorestframework==3.13.1
drf-yasg==1.21.3
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
packaging==21.3
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.8
zope.interface==5.4.0
